chore : CoupleAniversaryResponse Mapper 함수 패키지 위치 변경
