+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/mapper/CoupleMapper.kt
@@ -55,42 +55,3 @@ internal fun CoupleBasicResponse.toCouple(): Couple =
         sharedMessage = this.sharedMessage ?: "",
         status = CoupleStatus.valueOf(this.status.name),
     )
-
-internal fun CoupleAnniversaryResponse.toAnniversary(): List<Anniversary> {
-    val hundredDayAnniversaries =
-        this.hundredDayAnniversaries.map {
-            Anniversary(
-                date = LocalDate.parse(it.date),
-                type = AnniversaryType.valueOf(it.type),
-                label = it.label,
-            )
-        }
-
-    val yearlyAnniversaries =
-        this.yearlyAnniversaries.map {
-            Anniversary(
-                date = LocalDate.parse(it.date),
-                type = AnniversaryType.valueOf(it.type),
-                label = it.label,
-            )
-        }
-
-    val myBirthdayAnniversaries =
-        this.myBirthDates.map {
-            Anniversary(
-                date = LocalDate.parse(it.date),
-                type = AnniversaryType.valueOf(it.type),
-                label = it.label,
-            )
-        }
-
-    val partnerBirthdayAnniversaries =
-        this.partnerBirthDates.map {
-            Anniversary(
-                date = LocalDate.parse(it.date),
-                type = AnniversaryType.valueOf(it.type),
-                label = it.label,
-            )
-        }
-    return hundredDayAnniversaries + yearlyAnniversaries + myBirthdayAnniversaries + partnerBirthdayAnniversaries
-}
